pyproject.toml
[project]
name = "ai-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "accelerate>=1.13.0",
    "datasets>=4.8.4",
    "python-dotenv>=1.2.2",
    "soundfile>=0.13.1",
    "torch>=2.11.0",
    "torchcodec>=0.11.0",
    "transformers>=5.5.0",
]
